Stricter approach to many-to-many JSON encoding.
The previous solution is simple, but more lenient in the case of inconsistent models. We should have automated verification to ensure every relationship is defined to prevent circular encoding or missing values.

The simple fix of checking whether the inverse relationship allows inverse encoding is both convoluted and too lenient. We have an implicit rule in our model that any relationship should specify ownership via implication of delete rules: if a relationship deletes its destination, then it owns it. If it owns it, the inverse should not encode itself, because it would be encoding its owner. But if a relationship's ownership cannot be inferred, then we also will encode from either direction, which could lead to a circular (recursive) encoding, as the same relationship is traversed back and forth.

There should not be any relationships like this in our model. Every relationship should have an owner. But we need to verify, and assert this assumption.