Update test_pod_version script
The git environment in the action runner is different than the type of
git environment we use locally. The runner action for checkout is
efficient in only getting the current version, so extra configuration is
necessary to do a diff.

Needed to make sure master exists, and reference it using the remote
branch.